requirements.txt
# 网络请求
httpx>=0.25.0
aiohttp>=3.9.0

# 网页解析
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0

# Gemini API
google-generativeai>=0.3.0

# 调度
apscheduler>=3.10.0

# 配置和日志
pyyaml>=6.0
loguru>=0.7.0

# 工具
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.31.0

# 开发测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
